serverless/apis/recommend/requirements.txt
opensearch-py==2.7.1
requests-aws4auth==1.3.1
boto3==1.35.57
google-cloud-bigquery==3.11.4
aws-xray-sdk==2.14.0
pandas==2.0.3
numpy==1.24.4